Refactor database module into a subpackage
